+++ home-assistant-addon/CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added a script to create a database containing demo data
 - Improved due date checking for consistency (overdue, issue violates release date, task violates issue deadline) and added appropriate style / notication
 - Removed edit button for users and projects in system settings, edit dialog is now opened by clicking the card
+- Bumped dompurify to 3.4.3
 
 ## 1.2.1